Move tests to `CardReaderPaymentController`
